Moved SocialUser class to its own file
Also simplify the class as a minimal helper
